# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Maverick Execution Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # USERS
...
  /assessments:
    get:
      operationId: listAssessments
      tags: [assessments]
      summary: List assessments
      parameters:
        - in: query
          name: batchId
          schema:
            type: integer
        - in: query
          name: type
          schema:
            type: string
      responses:
        "200":
          description: List of assessments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assessment"
    post:
      operationId: createAssessment
      tags: [assessments]
      summary: Create an assessment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "201":
          description: Created assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
  /assessments/{id}:
    get:
      operationId: getAssessment
      tags: [assessments]
      summary: Get an assessment by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
        "404":
          description: Not found
    patch:
      operationId: updateAssessment
      tags: [assessments]
      summary: Update an assessment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentUpdate"
      responses:
        "200":
          description: Updated assessment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assessment"
    delete:
      operationId: deleteAssessment
      tags: [assessments]
      summary: Delete an assessment
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /assessment-scores:
    get:
      operationId: listAssessmentScores
      tags: [assessments]
      summary: List assessment scores
      parameters:
        - in: query
          name: assessmentId
          schema:
            type: integer
        - in: query
          name: candidateId
          schema:
            type: integer
        - in: query
          name: batchId
          schema:
            type: integer
      responses:
        "200":
          description: List of scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentScore"
    post:
      operationId: createAssessmentScore
      tags: [assessments]
      summary: Submit an assessment score
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentScoreInput"
      responses:
        "201":
          description: Created score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentScore"
  /assessment-scores/bulk:
    post:
      operationId: bulkCreateAssessmentScores
      tags: [assessments]
      summary: Submit bulk assessment scores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAssessmentScoreInput"
      responses:
        "201":
          description: Created scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AssessmentScore"
  /assessment-scores/{id}:
    patch:
      operationId: updateAssessmentScore
      tags: [assessments]
      summary: Update an assessment score
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentScoreUpdate"
      responses:
        "200":
          description: Updated score
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentScore"

  # TOPPERS
...
  /assessments/bulk-scores:
    post:
      operationId: bulkUploadAssessmentScores
      tags: [assessments]
      summary: Bulk upload assessment scores
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUploadScoresBody"
      responses:
        "200":
          description: Upload result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkImportResult"

  # REPORTS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"

    # USERS
    User:
      type: object
      required: [id, name, email, role, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, coordinator, trainer]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
...
    Assessment:
      type: object
      required: [id, batchId, title, type, scheduledDate, maxScore, createdAt]
      properties:
        id:
          type: integer
        batchId:
          type: integer
        batchName:
          type: ["string", "null"]
        title:
          type: string
        type:
          type: string
          enum: [sprint_review, coding, api, project_evaluation]
        scheduledDate:
          type: string
          format: date
        maxScore:
          type: number
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    AssessmentInput:
      type: object
      required: [batchId, title, type, scheduledDate, maxScore]
      properties:
        batchId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [sprint_review, coding, api, project_evaluation]
        scheduledDate:
          type: string
          format: date
        maxScore:
          type: number
        description:
          type: string
    AssessmentUpdate:
      type: object
      properties:
        title:
          type: string
        scheduledDate:
          type: string
          format: date
        maxScore:
          type: number
        description:
          type: string
    AssessmentScore:
      type: object
      required: [id, assessmentId, candidateId, score, createdAt]
      properties:
        id:
          type: integer
        assessmentId:
          type: integer
        assessmentTitle:
          type: ["string", "null"]
        assessmentType:
          type: ["string", "null"]
        candidateId:
          type: integer
        candidateName:
          type: ["string", "null"]
        score:
          type: number
        maxScore:
          type: ["number", "null"]
        percentage:
          type: ["number", "null"]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    AssessmentScoreInput:
      type: object
      required: [assessmentId, candidateId, score]
      properties:
        assessmentId:
          type: integer
        candidateId:
          type: integer
        score:
          type: number
        remarks:
          type: string
    AssessmentScoreUpdate:
      type: object
      properties:
        score:
          type: number
        remarks:
          type: string
    BulkAssessmentScoreInput:
      type: object
      required: [assessmentId, scores]
      properties:
        assessmentId:
          type: integer
        scores:
          type: array
          items:
            type: object
            required: [candidateId, score]
            properties:
              candidateId:
                type: integer
              score:
                type: number
              remarks:
                type: string

    # TOPPERS
...
    BulkUploadScoresBody:
      type: object
      required: [assessmentId, scores]
      properties:
        assessmentId:
          type: integer
        scores:
          type: array
          items:
            type: object
            required: [candidateId, score]
            properties:
              candidateId:
                type: integer
              score:
                type: number
              remarks:
                type: string
    BulkImportResult:
      type: object
      required: [inserted, failed, errors]
      properties:
        inserted:
          type: integer
        failed:
          type: integer
        errors:
          type: array
          items:
            type: string
...